completes animated hangman on home page
